requirements.txt
aiodns==3.2.0
asyncio==3.4.3
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==2.0.12
colorama==0.4.5
dnspython==2.2.1
idna==3.6
importlib-resources==5.4.0
mysql-connector-python==9.0.0
numpy==1.24.4
pandas==2.0.3
pika==1.3.1
pycares==4.4.0
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.27.1
schedule==1.1.0
six==1.16.0
tqdm==4.64.1
tzdata==2024.2
urllib3==1.26.18
uvloop==0.20.0
zipp==3.6.0
